:recycle: Improve code structure encoder => protocol
