+++ lib/elven_gard/protocol/binary/integer_type.ex
@@ -1,24 +1,24 @@
-defmodule ElvenGard.Types.Binary.LongType do
+defmodule ElvenGard.Protocol.Binary.IntegerType do
   @moduledoc """
-  Define a custom long type (uint64_t) for game protocols
+  Define a custom integer type (uint32_t) for game protocols
   """
 
   use ElvenGard.Helpers.Type
 
   @impl ElvenGard.Helpers.Type
   @spec encode(integer, list) :: bitstring
-  def encode(long, _opts) do
-    <<long::little-size(64)>>
+  def encode(int, _opts) do
+    <<int::little-size(32)>>
   end
 
   @impl ElvenGard.Helpers.Type
   @spec decode(bitstring, list) :: {integer, bitstring}
   def decode(bin, _opts) do
     <<
-      long::little-size(64),
+      int::little-size(32),
       rest::binary
     >> = bin
 
-    {long, rest}
+    {int, rest}
   end
 end
